# src/main/java/com/example/chatonymous/users/controller/UsersController.java
package com.example.chatonymous.users.controller;

import com.example.chatonymous.users.model.JwtResponceRecord;
import com.example.chatonymous.users.model.UserNamePasswordRecord;
import com.example.chatonymous.users.services.TokenService;
import com.example.chatonymous.users.model.UserModel;
import com.example.chatonymous.users.repository.UsersRepository;
import com.example.chatonymous.users.services.UsersServices;
import lombok.AllArgsConstructor;
import org.springframework.http.ResponseEntity;
import org.springframework.security.authentication.AuthenticationManager;
import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
import org.springframework.web.bind.annotation.*;

import java.util.List;

@CrossOrigin(origins = "*")
@AllArgsConstructor
@RestController
@RequestMapping("/api/users")
public class UsersController {
    private UsersServices usersServices;
    private AuthenticationManager authManager;
    private TokenService tokenService;

    @GetMapping
    public ResponseEntity<List<UserModel>> getMapping(@RequestParam(required = false, value = "username") String username) {
        return usersServices.findByOrUserName(username);
    }

    @PostMapping("/signup")
    public ResponseEntity<?> signUp(@RequestBody UserNamePasswordRecord userRecord) {
        return usersServices.signup(userRecord);
    }

    @PostMapping("/login")
    public ResponseEntity<JwtResponceRecord> login(@RequestBody UserNamePasswordRecord userRecord) {
        var auth = authManager.authenticate(
                new UsernamePasswordAuthenticationToken(userRecord.userName(), userRecord.passWord())
        );

        String token = tokenService.generate(auth.getName());
        return ResponseEntity.ok(new JwtResponceRecord(token));
    }

    @DeleteMapping("/{userName}")
    public ResponseEntity<?> deleteMapping(@PathVariable String userName) {
        return usersServices.delete(userName);
    }
}